+++ mlflow/server/js/src/experiment-tracking/components/ExperimentView.test.js
@@ -29,6 +29,7 @@ import {
   DEFAULT_START_TIME,
   COLUMN_SORT_BY_ASC,
   COLUMN_SORT_BY_DESC,
+  ATTRIBUTE_COLUMN_LABELS,
 } from '../constants';
 
 let onSearchSpy;
@@ -437,5 +438,116 @@ describe('Start time dropdown', () => {
       MODEL_VERSION_FILTER.ALL_RUNS,
       'LAST_7_DAYS',
     );
+  });
+});
+
+describe('Diff Switch', () => {
+  test('handleDiffSwitchChange changes state correctly', () => {
+    const getCategorizedColumnsDiffViewSpy = jest.fn();
+    const handleColumnSelectionCheckSpy = jest.fn();
+    const wrapper = getExperimentViewMock();
+    const instance = wrapper.instance();
+    instance.getCategorizedColumnsDiffView = getCategorizedColumnsDiffViewSpy;
+    instance.handleColumnSelectionCheck = handleColumnSelectionCheckSpy;
+
+    // Checkbox unmarked by default
+    expect(wrapper.state().diffSwitchSelected).toBe(false);
+
+    // Checkbox marked
+    instance.handleDiffSwitchChange();
+    expect(wrapper.state().diffSwitchSelected).toBe(true);
+    expect(getCategorizedColumnsDiffViewSpy).toHaveBeenCalledTimes(1);
+    expect(handleColumnSelectionCheckSpy).toHaveBeenCalledTimes(1);
+
+    // Checkbox unmarked
+    instance.handleDiffSwitchChange();
+    expect(wrapper.state().diffSwitchSelected).toBe(false);
+    expect(getCategorizedColumnsDiffViewSpy).toHaveBeenCalledTimes(1);
+    expect(handleColumnSelectionCheckSpy).toHaveBeenCalledTimes(2);
+    expect(handleColumnSelectionCheckSpy).toHaveBeenLastCalledWith({
+      [COLUMN_TYPES.ATTRIBUTES]: [],
+      [COLUMN_TYPES.PARAMS]: [],
+      [COLUMN_TYPES.METRICS]: [],
+      [COLUMN_TYPES.TAGS]: [],
+    });
+  });
+
+  test('getCategorizedColumnsDiffView returns the correct column keys to uncheck', () => {
+    const runInfos = [
+      RunInfo.fromJs({
+        run_uuid: 'run-id1',
+        experiment_id: '3',
+        status: 'FINISHED',
+        start_time: 1,
+        end_time: 1,
+        artifact_uri: 'dummypath',
+        lifecycle_stage: 'active',
+      }),
+      RunInfo.fromJs({
+        run_uuid: 'run-id2',
+        experiment_id: '3',
+        status: 'FINISHED',
+        start_time: 2,
+        end_time: 2,
+        artifact_uri: 'dummypath',
+        lifecycle_stage: 'active',
+      }),
+    ];
+    const paramKeyList = ['param1', 'param2', 'param3', 'param4'];
+    const metricKeyList = ['metric1', 'metric2', 'metric3', 'metric4'];
+    const paramsList = [
+      [
+        Param.fromJs({ key: 'param1', value: '1' }),
+        Param.fromJs({ key: 'param2', value: '1' }),
+        Param.fromJs({ key: 'param3', value: '1' }),
+      ],
+      [Param.fromJs({ key: 'param1', value: '1' }), Param.fromJs({ key: 'param2', value: '2' })],
+    ];
+    const metricsList = [
+      [
+        Metric.fromJs({ key: 'metric1', value: '1' }),
+        Metric.fromJs({ key: 'metric2', value: '1' }),
+        Metric.fromJs({ key: 'metric3', value: '1' }),
+      ],
+      [
+        Metric.fromJs({ key: 'metric1', value: '1' }),
+        Metric.fromJs({ key: 'metric2', value: '2' }),
+      ],
+    ];
+    const tagsList = [
+      createTags({
+        tag1: '1',
+        tag2: '1',
+        tag3: '1',
+        [Utils.runNameTag]: 'runname1',
+        [Utils.userTag]: 'usertag1',
+        [Utils.sourceNameTag]: 'sourcename1',
+        [Utils.gitCommitTag]: 'gitcommit1',
+      }),
+      createTags({
+        tag1: '1',
+        tag2: '2',
+        [Utils.runNameTag]: 'runname1',
+        [Utils.userTag]: 'usertag2',
+        [Utils.sourceNameTag]: 'sourcename1',
+        [Utils.gitCommitTag]: 'gitcommit2',
+      }),
+    ];
+    const expectedUncheckedKeys = {
+      [COLUMN_TYPES.ATTRIBUTES]: [ATTRIBUTE_COLUMN_LABELS.RUN_NAME, ATTRIBUTE_COLUMN_LABELS.SOURCE],
+      [COLUMN_TYPES.PARAMS]: ['param1', 'param4'],
+      [COLUMN_TYPES.METRICS]: ['metric1', 'metric4'],
+      [COLUMN_TYPES.TAGS]: ['tag1'],
+    };
+    const wrapper = getExperimentViewMock({
+      runInfos,
+      paramKeyList,
+      metricKeyList,
+      paramsList,
+      metricsList,
+      tagsList,
+    });
+    const instance = wrapper.instance();
+    expect(instance.getCategorizedColumnsDiffView()).toEqual(expectedUncheckedKeys);
   });
 });
