+++ mlflow/server/js/src/experiment-tracking/components/RunsTableColumnSelectionDropdown.test.js
@@ -1,9 +1,7 @@
 import React from 'react';
 import { RunsTableColumnSelectionDropdown } from './RunsTableColumnSelectionDropdown';
 import { SearchTree } from '../../common/components/SearchTree';
-import { Metric, Param, RunInfo, RunTag } from '../sdk/MlflowMessages';
-import Utils from '../../common/utils/Utils';
-import { COLUMN_TYPES, ATTRIBUTE_COLUMN_LABELS } from '../constants';
+import { COLUMN_TYPES } from '../constants';
 import { mountWithIntl, shallowWithIntl } from '../../common/utils/TestUtils';
 
 describe('RunsTableColumnSelectionDropdown', () => {
@@ -122,117 +120,5 @@ describe('RunsTableColumnSelectionDropdown', () => {
       'metrics-m2',
       'tags-t2',
     ]);
-  });
-
-  test('handleDiffViewCheckboxChange changes state correctly', () => {
-    const getCategorizedColumnsDiffViewSpy = jest.fn();
-
-    wrapper = mountWithIntl(<RunsTableColumnSelectionDropdown {...commonProps} />);
-    instance = wrapper.instance();
-    instance.getCategorizedColumnsDiffView = getCategorizedColumnsDiffViewSpy;
-
-    // Checkbox unmarked by default
-    expect(wrapper.state().diffViewSelected).toBe(false);
-
-    // Checkbox marked
-    instance.handleDiffViewCheckboxChange();
-    expect(wrapper.state().diffViewSelected).toBe(true);
-    expect(getCategorizedColumnsDiffViewSpy).toHaveBeenCalledTimes(1);
-    expect(instance.props.onCheck).toHaveBeenCalledTimes(1);
-
-    // Checkbox unmarked
-    instance.handleDiffViewCheckboxChange();
-    expect(wrapper.state().diffViewSelected).toBe(false);
-    expect(getCategorizedColumnsDiffViewSpy).toHaveBeenCalledTimes(1);
-    expect(instance.props.onCheck).toHaveBeenCalledTimes(2);
-    expect(instance.props.onCheck).toHaveBeenLastCalledWith({
-      [COLUMN_TYPES.ATTRIBUTES]: [],
-      [COLUMN_TYPES.PARAMS]: [],
-      [COLUMN_TYPES.METRICS]: [],
-      [COLUMN_TYPES.TAGS]: [],
-    });
-  });
-
-  test('getCategorizedColumnsDiffView returns the correct column keys to uncheck', () => {
-    const runInfos = [
-      RunInfo.fromJs({
-        run_uuid: 'run-id1',
-        experiment_id: '3',
-        status: 'FINISHED',
-        start_time: 1,
-        end_time: 1,
-        artifact_uri: 'dummypath',
-        lifecycle_stage: 'active',
-      }),
-      RunInfo.fromJs({
-        run_uuid: 'run-id2',
-        experiment_id: '3',
-        status: 'FINISHED',
-        start_time: 2,
-        end_time: 2,
-        artifact_uri: 'dummypath',
-        lifecycle_stage: 'active',
-      }),
-    ];
-    const paramKeyList = ['param1', 'param2', 'param3', 'param4'];
-    const metricKeyList = ['metric1', 'metric2', 'metric3', 'metric4'];
-    const visibleTagKeyList = ['tag1', 'tag2', 'tag3'];
-    const paramsList = [
-      [
-        Param.fromJs({ key: 'param1', value: '1' }),
-        Param.fromJs({ key: 'param2', value: '1' }),
-        Param.fromJs({ key: 'param3', value: '1' }),
-      ],
-      [Param.fromJs({ key: 'param1', value: '1' }), Param.fromJs({ key: 'param2', value: '2' })],
-    ];
-    const metricsList = [
-      [
-        Metric.fromJs({ key: 'metric1', value: '1' }),
-        Metric.fromJs({ key: 'metric2', value: '1' }),
-        Metric.fromJs({ key: 'metric3', value: '1' }),
-      ],
-      [
-        Metric.fromJs({ key: 'metric1', value: '1' }),
-        Metric.fromJs({ key: 'metric2', value: '2' }),
-      ],
-    ];
-    const tagsList = [
-      {
-        tag1: RunTag.fromJs({ key: 'tag1', value: '1' }),
-        tag2: RunTag.fromJs({ key: 'tag2', value: '1' }),
-        tag3: RunTag.fromJs({ key: 'tag3', value: '1' }),
-        [Utils.runNameTag]: RunTag.fromJs({ key: [Utils.runNameTag], value: 'runname1' }),
-        [Utils.userTag]: RunTag.fromJs({ key: [Utils.userTag], value: 'usertag1' }),
-        [Utils.sourceNameTag]: RunTag.fromJs({ key: [Utils.sourceNameTag], value: 'sourcename1' }),
-        [Utils.gitCommitTag]: RunTag.fromJs({ key: [Utils.gitCommitTag], value: 'gitcommit1' }),
-      },
-      {
-        tag1: RunTag.fromJs({ key: 'tag1', value: '1' }),
-        tag2: RunTag.fromJs({ key: 'tag2', value: '2' }),
-        [Utils.runNameTag]: RunTag.fromJs({ key: [Utils.runNameTag], value: 'runname1' }),
-        [Utils.userTag]: RunTag.fromJs({ key: [Utils.userTag], value: 'usertag2' }),
-        [Utils.sourceNameTag]: RunTag.fromJs({ key: [Utils.sourceNameTag], value: 'sourcename1' }),
-        [Utils.gitCommitTag]: RunTag.fromJs({ key: [Utils.gitCommitTag], value: 'gitcommit2' }),
-      },
-    ];
-    const expectedUncheckedKeys = {
-      [COLUMN_TYPES.ATTRIBUTES]: [ATTRIBUTE_COLUMN_LABELS.RUN_NAME, ATTRIBUTE_COLUMN_LABELS.SOURCE],
-      [COLUMN_TYPES.PARAMS]: ['param1', 'param4'],
-      [COLUMN_TYPES.METRICS]: ['metric1', 'metric4'],
-      [COLUMN_TYPES.TAGS]: ['tag1'],
-    };
-    const props = {
-      ...minimalProps,
-      runInfos,
-      paramKeyList,
-      metricKeyList,
-      visibleTagKeyList,
-      paramsList,
-      metricsList,
-      tagsList,
-    };
-    wrapper = mountWithIntl(<RunsTableColumnSelectionDropdown {...props} />);
-    instance = wrapper.instance();
-    expect(instance.getCategorizedColumnsDiffView()).toEqual(expectedUncheckedKeys);
   });
 });
